==> notebooks/MachineLearning/1_SignatureExtracts.ipynb <==
# Signature Extracts

Create 5-day averaged CARD-BS signatures for all fields in the dk2018_bs_signatures.


from datetime import datetime, timedelta
import psycopg2


# Conect to the database
conn = psycopg2.connect(
    host="localhost",
    database="postgres",
    user="postgres",
    password="")
cur = conn.cursor()


# Set the table names
parcels_table = "aoi2020"
bs_extracts = 'aoi2020_bs_extracts'
bs_signatures = 'aoi2020_bs_signatures'
parcels_1 = 'aoi2020'

# Set stert and end date
start_date = datetime.strptime("2018-04-16", '%Y-%m-%d')
end_date = datetime.strptime("2018-12-27", '%Y-%m-%d')


step_period = 5
start_period = 0

start_date = start_date + timedelta(days=start_period*step_period)

insertSql = f"""
    INSERT into {bs_extracts} (pid, band, period, mean, count) 
    SELECT pid::int, band, {{}} As period, avg(mean) As mean, count(mean) As count 
    FROM {bs_signatures}, {parcels_table} 
    WHERE obsid In (SELECT id from {dias_catalog}
        WHERE status ='extracted' And card ='bs'
        And obstime between '{start_date}' And '{end_date}')
    And pid = {parcels_table}.ogc_fid
    And inaoi
    GROUP by pid, band;
"""

for i in range(start_period, 5):
    end_date = start_date + timedelta(days=step_period)
    print(insertSql)
    try:
        cur.execute(insertSql.format(i))
        conn.commit()
    except Error as e:
        print(e)
    
    print(i, start_date, end_date)
    
    start_date = end_date

# Close database connection 
database.close_connection()
