==> notebooks/MachineLearning/3_PrepareFeatureSet.ipynb <==
## PrepareFeatureSet

In order to run machine learning algorithms, we need to prepare feature vectors (tensors) that will be signature extracts.
For this, we need to reformat the data in the data base so that for each parcel, we have the full signature in one row.

We save the result to .csv 

import psycopg2
import pandas as pd

# Conect to the database
conn = psycopg2.connect(
    host="localhost",
    database="postgres",
    user="postgres",
    password="")
cur = conn.cursor()

# Set the table names
parcels_table = "aoi2020"
crop_type_column = "" # Set the crop type column

bs_signatures = "aoi2020_bs_signatures"
s2_signatures = "aoi2020_s2_signatures"
c6_signatures = "aoi2020_c6_signatures"
bs_extracts = "aoi2020_bs_extracts"
date_start = "2019-01-01"
date_end = "2019-10-15"


# Set the folder to store the data
data_folder = ''


selectSql = f"""
    SELECT e1.pid::int as pid, e1.period as period, e1.mean as vv, e2.mean as vh
    FROM {bs_extracts} e1, {bs_extracts} e2
    WHERE e1.pid = e2.pid
    And e1.period = e2.period
    And e1.band = 'VV'
    And e2.band = 'VH';
    """

df = pd.read_sql_query(selectSql, conn)
cur.close()
conn.close()

df[0:10]

df.set_index('pid', inplace=True)

len(df)

periods = sorted(df['period'].unique())

periods

startperiod = 11
endperiod = 54

p0 = df[df['period']==startperiod].copy()
p0.drop(['period'], axis=1, inplace=True)
p0.columns = [f'vv_{startperiod}', f'vh_{startperiod}']

for p in periods[startperiod+1:endperiod]:
    pN = df[df['period']==p].copy()
    pN.drop(['period'], axis=1, inplace=True)
    pN.columns = [f'vv_{p}', f'vh_{p}']

    p0 = p0.join(pN)

p0[0:10]

p0.to_csv(f"{data_}{parcels_table}_bs_features.csv")

# Close database connection
database.close_connection()
