==> notebooks/MachineLearning/4_CreateClassFeatures.ipynb <==
## Create calss features

import pandas as pd
import numpy as np

# Set the table names
parcels_table = "aoi2020"

# Set the folder to store the data
data_folder = ''

dfs = pd.read_csv(f'{data_}{parcels_table}_bs_features.csv')
dcc = pd.read_csv(f'{data_}{parcels_table}_cropclasses.csv')

dcombi = pd.merge(dfs, dcc, left_on='pid', right_on='ogc_fid')
dcombi.drop('ogc_fid', axis=1, inplace=True)

dcombi.dropna(inplace=True)

print(len(dcombi))
print(dcombi[0:10])


chunks = 4

df_test = dcombi.copy(deep=True)

samplesize = len(dcombi)//chunks
print(samplesize)

for i in range(chunks-1):
    print(i, len(dcombi))
    training = dcombi.take(np.random.permutation(len(dcombi)))[:samplesize]

    training.to_csv(f'{data_}{parcels_table}_train_{i}.csv')

    dcombi.drop(training.index, inplace=True)

    df_test.loc[~df_test.index.isin(training.index)].to_csv(f'{data_}{parcels_table}_test_{i}.csv')


print ("Remainder", len(dcombi))
dcombi.to_csv(f'{data_}{parcels_table}_train_{chunks-1}.csv')
df_test.loc[~df_test.index.isin(dcombi.index)].to_csv(f'{data_}{parcels_table}_test_{chunks-1}.csv')
