==> notebooks/MachineLearning/8_OverallConfusionMatrix.ipynb <==
# Get the overall confusion matrix

This notebook takes the aggregated predictions generated in the previous notebook and arranges them in a confusion matrix. We make the criterion for comission/omission quite strict by ensuring that an outlier is defined only if all 3 predicted class labels are the same and NOT EQUAL to the labeled class. Relaxing this rule will result in more confusion.

We show results in terms of parcel counts and in terms of total area in the parcels. Not surprisingly, the overall accuracy in the latter is significantly larger, as most confusion occurs with smaller parcels.



import psycopg2
import pandas as pd
import numpy as np
import psycopg2
from collections import Counter

# Conect to the database
conn = psycopg2.connect(
    host="localhost",
    database="postgres",
    user="postgres",
    password="")
cur = conn.cursor()

# Set the table names
parcels_table = "aoi2020"
bs_tensorflow = "bs_tensorflow"

# Set the folder to store the data
data_folder = ''


classLabelSql = F"""
    SELECT pid::int, class::int, majclass::int, majcount::int,
        st_area(wkb_geometry)/10000.0 as area 
    FROM {bs_tensorflow} tf, {parcels_table}
    WHERE tf.pid = {parcels}.ogc_fid
    """

df = pd.read_sql_query(classLabelSql, conn)

carea = np.zeros((11, 11))
ccnt = np.zeros((11, 11))

for i in df.index:
    row = df.loc[i]
    n = int(row['class'])    
    m = int(row['majclass'])
    area = float(row['area'])
    cnt = int(row['majcount'])
    if (cnt > 2) and (n != m):
        carea[n,m] += area
        ccnt[n,m] += 1
    else:
        carea[n,n] += area
        ccnt[n,n] += 1

np.set_printoptions(suppress=True, precision=0)
pd.set_option('expand_frame_repr', False)
pd.set_option('precision', 1)

print ("Overall Accuracy (area): ", 100.0*carea.trace()/carea.sum())

# These are the ordered labels for class 0 to 10
crops =['GRA', 'MAI', 'POT', 'WWH', 'SBT', 'WBA', 'WOR', 'SCE', 'WCE', 'VEG', 'FAL']

print(pd.DataFrame(carea, index = crops, columns = crops))

print ("\n\nOverall Accuracy (count): ", 100.0*ccnt.trace()/ccnt.sum())

print(pd.DataFrame(ccnt, index = crops, columns = crops))



You can now move back to the general notebook session to select and plot the outlier signatures versus their neighboring conformist parcels.


# Close database connection
database.close_connection()
